requirements.txt
dspy_ai==2.4.5
Requests==2.32.3
